fix: add bookkeeping info to correlate deprecated init() with module
This workaround keeps track of all the calls to `napi_module_register()` and collects some bookkeeping data to find which module (loaded addon) it belongs to. Currently, it tries to find the initializer function by symbol name (which was looked up via `dladdr`), assuming that the symbol for this function was exported.

NOTE: The "registry" of such deprecated addons MUST be static, as `AddonRegistry::handleOldNapiModuleRegister()` will be called BEFORE `AddonRegistry`'s constructor. Without it, the registry will be emptied during "construction" and the mutex won't be initialized leaving you with a SIGABRT: `std::__1::system_error: mutex lock failed: Invalid argument`