Cherry pick the `findNodeAddonForBindings()` to path-utils (#152)
* feat: cherry-pick findNodeAddonForBindings()

* feat: relax the condition for CJS modules without file exts

This small change relaxes the condition for taking the shortcut, as CommonJS modules (in contrast to ESM) do not require developers to explicitly include the file extensions. The Node.js module resolution algorithm (https://nodejs.org/api/modules.html#all-together) in step 4 of LOAD_AS_FILE(X) would try appending the `.node` extension.

In theory, we should make sure that other extensions checked in previous steps are not present, but given that we are implementing it for `requireNodeAddon()`, it should be safe to skip those.

# Conflicts:
#	packages/host/src/node/path-utils.ts

* refactor: check each path in separate tests

* style: fix eslint issues

---------

Co-authored-by: Mariusz Pasinski <mariusz.pasinski@callstack.com>